Initial Stages Complete:
At this stage in the program, the website functions properly with registering and login features added as well. The semester and module models have been impletemented, with surface level restrictions added to logged in members only being able to view the details.
The website links to an azure sql database that is updated via an SSMS connection.

Furture updates will include: restrictions to logged in user-only information shown, Study guide tracker, website customization, and publishing the site publically.